+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/altar/AncientAltar.java
@@ -1,5 +1,8 @@
 package io.github.thebusybiscuit.slimefun4.implementation.items.altar;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.slimefun4.api.events.AncientAltarCraftEvent;
@@ -26,6 +29,7 @@ import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 public class AncientAltar extends SlimefunItem {
 
     private final int speed;
+    private final List<AltarRecipe> recipes = new ArrayList<>();
 
     public AncientAltar(Category category, int speed, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe) {
         super(category, item, recipeType, recipe);
@@ -46,6 +50,10 @@ public class AncientAltar extends SlimefunItem {
      */
     public int getSpeed() {
         return speed;
+    }
+
+    public List<AltarRecipe> getRecipes() {
+        return recipes;
     }
 
 }