+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/AncientAltarListener.java
@@ -8,14 +8,12 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 
-import org.bukkit.ChatColor;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Sound;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
-import org.bukkit.entity.Entity;
 import org.bukkit.entity.Item;
 import org.bukkit.entity.Player;
 import org.bukkit.event.Event.Result;
@@ -25,9 +23,7 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
-import org.bukkit.util.Vector;
 
-import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
 import io.github.thebusybiscuit.cscorelib2.inventory.ItemUtils;
 import io.github.thebusybiscuit.cscorelib2.item.CustomItem;
 import io.github.thebusybiscuit.cscorelib2.protection.ProtectableAction;
@@ -54,19 +50,19 @@ import me.mrCookieSlime.Slimefun.api.Slimefun;
  */
 public class AncientAltarListener implements Listener {
 
-    public static final String ITEM_PREFIX = ChatColors.color("&dALTAR &3Probe - &e");
+    private AncientAltar altarItem;
+    private AncientPedestal pedestalItem;
 
-    private AncientAltar altar;
-
-    private final Set<AltarRecipe> altarRecipes = new HashSet<>();
     private final Set<Location> altarsInUse = new HashSet<>();
 
     private final List<Block> altars = new ArrayList<>();
     private final Set<UUID> removedItems = new HashSet<>();
 
-    public void register(SlimefunPlugin plugin, AncientAltar altar) {
+    public AncientAltarListener(SlimefunPlugin plugin, AncientAltar altar, AncientPedestal pedestal) {
         plugin.getServer().getPluginManager().registerEvents(this, plugin);
-        this.altar = altar;
+
+        this.altarItem = altar;
+        this.pedestalItem = pedestal;
     }
 
     /**
@@ -82,13 +78,9 @@ public class AncientAltarListener implements Listener {
         return altars;
     }
 
-    public Set<AltarRecipe> getRecipes() {
-        return altarRecipes;
-    }
-
     @EventHandler
     public void onInteract(PlayerRightClickEvent e) {
-        if (altar == null || altar.isDisabled() || e.useBlock() == Result.DENY) {
+        if (altarItem == null || altarItem.isDisabled() || e.useBlock() == Result.DENY) {
             return;
         }
 
@@ -109,12 +101,12 @@ public class AncientAltarListener implements Listener {
 
         String id = slimefunBlock.get().getID();
 
-        if (id.equals(SlimefunItems.ANCIENT_PEDESTAL.getItemId())) {
+        if (id.equals(pedestalItem.getID())) {
             e.cancel();
             usePedestal(b, e.getPlayer());
         }
-        else if (id.equals(SlimefunItems.ANCIENT_ALTAR.getItemId())) {
-            if (!Slimefun.hasUnlocked(e.getPlayer(), SlimefunItems.ANCIENT_ALTAR.getItem(), true) || altarsInUse.contains(b.getLocation())) {
+        else if (id.equals(altarItem.getID())) {
+            if (!Slimefun.hasUnlocked(e.getPlayer(), altarItem, true) || altarsInUse.contains(b.getLocation())) {
                 e.cancel();
                 return;
             }
@@ -138,9 +130,9 @@ public class AncientAltarListener implements Listener {
         }
 
         // getting the currently placed item
-        Item stack = findItem(pedestal);
+        Optional<Item> stack = pedestalItem.getPlacedItem(pedestal);
 
-        if (stack == null) {
+        if (!stack.isPresent()) {
             // Check if the Item in hand is valid
             if (p.getInventory().getItemInMainHand().getType() != Material.AIR) {
                 // Check for pedestal obstructions
@@ -150,17 +142,18 @@ public class AncientAltarListener implements Listener {
                 }
 
                 // place the item onto the pedestal
-                insertItem(p, pedestal);
+                pedestalItem.placeItem(p, pedestal);
             }
         }
-        else if (!removedItems.contains(stack.getUniqueId())) {
-            UUID uuid = stack.getUniqueId();
+        else if (!removedItems.contains(stack.get().getUniqueId())) {
+            Item entity = stack.get();
+            UUID uuid = entity.getUniqueId();
             removedItems.add(uuid);
 
             Slimefun.runSync(() -> removedItems.remove(uuid), 30L);
 
-            stack.remove();
-            p.getInventory().addItem(fixItemStack(stack.getItemStack(), stack.getCustomName()));
+            entity.remove();
+            p.getInventory().addItem(pedestalItem.getOriginalItemStack(entity));
             p.playSound(pedestal.getLocation(), Sound.ENTITY_ITEM_PICKUP, 1F, 1F);
         }
     }
@@ -209,10 +202,10 @@ public class AncientAltarListener implements Listener {
         List<ItemStack> input = new ArrayList<>();
 
         for (Block pedestal : pedestals) {
-            Item stack = findItem(pedestal);
+            Optional<Item> stack = pedestalItem.getPlacedItem(pedestal);
 
-            if (stack != null) {
-                input.add(fixItemStack(stack.getItemStack(), stack.getCustomName()));
+            if (stack.isPresent()) {
+                input.add(pedestalItem.getOriginalItemStack(stack.get()));
             }
         }
 
@@ -227,7 +220,9 @@ public class AncientAltarListener implements Listener {
                 }
 
                 b.getWorld().playSound(b.getLocation(), Sound.ENTITY_ILLUSIONER_PREPARE_MIRROR, 1, 1);
-                Slimefun.runSync(new AncientAltarTask(b, altar.getSpeed(), result.get(), pedestals, consumed, p), 10L);
+
+                AncientAltarTask task = new AncientAltarTask(this, b, altarItem.getSpeed(), result.get(), pedestals, consumed, p);
+                Slimefun.runSync(task, 10L);
             }
             else {
                 altars.remove(b);
@@ -255,7 +250,7 @@ public class AncientAltarListener implements Listener {
 
     @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
     public void onBlockPlace(BlockPlaceEvent e) {
-        if (altar == null || altar.isDisabled()) {
+        if (altarItem == null || altarItem.isDisabled()) {
             return;
         }
 
@@ -264,82 +259,38 @@ public class AncientAltarListener implements Listener {
         if (pedestal.getType() == Material.DISPENSER) {
             String id = BlockStorage.checkID(pedestal);
 
-            if (id != null && id.equals(SlimefunItems.ANCIENT_PEDESTAL.getItemId())) {
+            if (id != null && id.equals(pedestalItem.getID())) {
                 SlimefunPlugin.getLocalization().sendMessage(e.getPlayer(), "messages.cannot-place", true);
                 e.setCancelled(true);
             }
         }
     }
 
-    public ItemStack fixItemStack(ItemStack itemStack, String customName) {
-        ItemStack stack = itemStack.clone();
-
-        if (customName.equals(ItemUtils.getItemName(new ItemStack(itemStack.getType())))) {
-            ItemMeta im = stack.getItemMeta();
-            im.setDisplayName(null);
-            stack.setItemMeta(im);
-        }
-        else {
-            ItemMeta im = stack.getItemMeta();
-            if (!customName.startsWith(String.valueOf(ChatColor.COLOR_CHAR))) customName = ChatColor.RESET + customName;
-            im.setDisplayName(customName);
-            stack.setItemMeta(im);
-        }
-        return stack;
-    }
-
-    public Item findItem(Block b) {
-        for (Entity n : b.getChunk().getEntities()) {
-            if (n instanceof Item && b.getLocation().add(0.5, 1.2, 0.5).distanceSquared(n.getLocation()) < 0.5D && n.getCustomName() != null) {
-                return (Item) n;
-            }
-        }
-        return null;
-    }
-
-    private void insertItem(Player p, Block b) {
-        ItemStack hand = p.getInventory().getItemInMainHand();
-        ItemStack stack = new CustomItem(hand, 1);
-
-        if (p.getGameMode() != GameMode.CREATIVE) {
-            ItemUtils.consumeItem(hand, false);
-        }
-
-        String nametag = ItemUtils.getItemName(stack);
-        Item entity = b.getWorld().dropItem(b.getLocation().add(0.5, 1.2, 0.5), new CustomItem(stack, ITEM_PREFIX + System.nanoTime()));
-        entity.setVelocity(new Vector(0, 0.1, 0));
-        SlimefunUtils.markAsNoPickup(entity, "altar_item");
-        entity.setCustomNameVisible(true);
-        entity.setCustomName(nametag);
-        p.playSound(b.getLocation(), Sound.ENTITY_ITEM_PICKUP, 0.3F, 0.3F);
-    }
-
     private List<Block> getPedestals(Block altar) {
         List<Block> list = new ArrayList<>();
-        String id = SlimefunItems.ANCIENT_PEDESTAL.getItemId();
 
-        if (BlockStorage.check(altar.getRelative(2, 0, -2), id)) {
+        if (BlockStorage.check(altar.getRelative(2, 0, -2), pedestalItem.getID())) {
             list.add(altar.getRelative(2, 0, -2));
         }
-        if (BlockStorage.check(altar.getRelative(3, 0, 0), id)) {
+        if (BlockStorage.check(altar.getRelative(3, 0, 0), pedestalItem.getID())) {
             list.add(altar.getRelative(3, 0, 0));
         }
-        if (BlockStorage.check(altar.getRelative(2, 0, 2), id)) {
+        if (BlockStorage.check(altar.getRelative(2, 0, 2), pedestalItem.getID())) {
             list.add(altar.getRelative(2, 0, 2));
         }
-        if (BlockStorage.check(altar.getRelative(0, 0, 3), id)) {
+        if (BlockStorage.check(altar.getRelative(0, 0, 3), pedestalItem.getID())) {
             list.add(altar.getRelative(0, 0, 3));
         }
-        if (BlockStorage.check(altar.getRelative(-2, 0, 2), id)) {
+        if (BlockStorage.check(altar.getRelative(-2, 0, 2), pedestalItem.getID())) {
             list.add(altar.getRelative(-2, 0, 2));
         }
-        if (BlockStorage.check(altar.getRelative(-3, 0, 0), id)) {
+        if (BlockStorage.check(altar.getRelative(-3, 0, 0), pedestalItem.getID())) {
             list.add(altar.getRelative(-3, 0, 0));
         }
-        if (BlockStorage.check(altar.getRelative(-2, 0, -2), id)) {
+        if (BlockStorage.check(altar.getRelative(-2, 0, -2), pedestalItem.getID())) {
             list.add(altar.getRelative(-2, 0, -2));
         }
-        if (BlockStorage.check(altar.getRelative(0, 0, -3), id)) {
+        if (BlockStorage.check(altar.getRelative(0, 0, -3), pedestalItem.getID())) {
             list.add(altar.getRelative(0, 0, -3));
         }
 
@@ -370,7 +321,7 @@ public class AncientAltarListener implements Listener {
     }
 
     private Optional<ItemStack> checkRecipe(ItemStack catalyst, List<ItemStackWrapper> items) {
-        for (AltarRecipe recipe : altarRecipes) {
+        for (AltarRecipe recipe : altarItem.getRecipes()) {
             if (SlimefunUtils.isItemSimilar(catalyst, recipe.getCatalyst(), true)) {
                 Optional<ItemStack> optional = checkPedestals(items, recipe);