+++ CHANGELOG.md
@@ -85,6 +85,12 @@
 * Fixed Electric Press not working
 * Fixed #2240
 * Fixed #2243
+* Fixed #2249
+* Fixed #1022
+* Fixed #2208
+* Fixed Fluid Pump treating low-level fluids like stationary fluids
+* Fixed Fluid Pump not working on Bubble Columns
+* Fixed #2251
 
 ## Release Candidate 15 (01 Aug 2020)